requirements.txt
fastapi==0.96.0
uvicorn[standard]==0.22.0
loguru==0.7.0
joblib==1.1.0
pandas==1.4.4
pytest==7.1.2
catboost==1.2.1
opentelemetry-api==1.19.0
opentelemetry-sdk==1.19.0
opentelemetry-instrumentation-asgi==0.40b0
opentelemetry-instrumentation-fastapi==0.40b0
opentelemetry-instrumentation-requests==0.40b0
opentelemetry-instrumentation-logging==0.40b0
opentelemetry-exporter-jaeger==1.19.0
opentelemetry-exporter-otlp-proto-grpc==1.19.0
opentelemetry-exporter-prometheus==1.12.0rc1
python-multipart==0.0.6
prometheus-client==0.17.1
requests==2.31.0